project multiple image

+++ app/Http/Controllers/ProjectController.php
@@ -18,9 +18,13 @@ class ProjectController extends Controller
      */
     public function index()
     {
-        return view('backend.pages.projects.index',[
-            'projects' => Project::latest()->paginate(10),
-        ]);
+        if(auth()->user()->can('project management')){
+            return view('backend.pages.projects.index',[
+                'projects' => Project::latest()->paginate(10),
+            ]);
+        }else{
+            return abort(404);
+        }
     }
 
     /**
@@ -30,7 +34,11 @@ class ProjectController extends Controller
      */
     public function create()
     {
-        return view('backend.pages.projects.create');
+        if(auth()->user()->can('project management')){
+            return view('backend.pages.projects.create');
+        }else{
+            return abort(404);
+        }
     }
 
     /**
@@ -41,26 +49,29 @@ class ProjectController extends Controller
      */
     public function store(Request $request)
     {
-        $request->validate([
-            'title'=> 'required|max:50',
-            'description'=> 'required|max:1000',
-            'image'=> 'required|mimes:jpg,jpeg,png,webp',
-        ]);
-        $project = new Project();
-        $project->title =$request->title;
-        $project->slug =Str::slug($request->title);
-        $project->description =$request->description;
-        $project->save();
-        if($request->hasFile('image')){
-            $image = $request->file('image');
-            $newName = Str::slug($project->title).'-image-'.Str::random(5).'.'.$image->getClientOriginalExtension();
-            $destination = public_path('images/projects/'.$newName);
-            Image::make($image)->save($destination);
-            $project->featured_image = $newName;
+        if(auth()->user()->can('project management')){
+            $request->validate([
+                'title'=> 'required|max:50',
+                'description'=> 'required|max:1000',
+                'image'=> 'required|mimes:jpg,jpeg,png,webp',
+            ]);
+            $project = new Project();
+            $project->title =$request->title;
+            $project->slug =Str::slug($request->title);
+            $project->description =$request->description;
             $project->save();
+            if($request->hasFile('image')){
+                $image = $request->file('image');
+                $newName = Str::slug($project->title).'-image-'.Str::random(5).'.'.$image->getClientOriginalExtension();
+                $destination = public_path('images/projects/'.$newName);
+                Image::make($image)->save($destination);
+                $project->featured_image = $newName;
+                $project->save();
+            }
+            return redirect()->route('projects.index')->with('success','Project Created!');
+        }else{
+            return abort(404);
         }
-        return redirect()->route('projects.index')->with('success','Project Created!');
-
     }
 
     /**
@@ -71,9 +82,13 @@ class ProjectController extends Controller
      */
     public function show(Project $project)
     {
-        return view('backend.pages.projects.show',[
-            'project' => $project,
-        ]);
+        if(auth()->user()->can('project management')){
+            return view('backend.pages.projects.show',[
+                'project' => $project,
+            ]);
+        }else{
+            return abort(404);
+        }
     }
 
     /**
@@ -84,9 +99,13 @@ class ProjectController extends Controller
      */
     public function edit(Project $project)
     {
-        return view('backend.pages.projects.edit',[
-            'project' => $project,
-        ]);
+        if(auth()->user()->can('project management')){
+            return view('backend.pages.projects.edit',[
+                'project' => $project,
+            ]);
+        }else{
+            return abort(404);
+        }
     }
 
     /**
@@ -98,34 +117,39 @@ class ProjectController extends Controller
      */
     public function update(Request $request, Project $project)
     {
-        $request->validate([
-            'title'=> 'required|max:50',
-            'description'=> 'required|max:1000',
-        ]);
-        if($request->hasFile('image')){
+        if(auth()->user()->can('project management')){
             $request->validate([
-                'image'=> 'required|mimes:png,jpg,webp,jpeg',
+                'title'=> 'required|max:50',
+                'description'=> 'required|max:1000',
             ]);
-        }
-        $project->title =$request->title;
-        $project->slug = Str::slug($request->title);
-        $project->description =$request->description;
-        $project->save();
-        if($request->hasFile('image')){
-            if($project->featured_image){
-                $oldImage = public_path('images/projects/'.$project->featured_image);
-                if(file_exists($oldImage)){
-                    unlink($oldImage);
-                }
+            if($request->hasFile('image')){
+                $request->validate([
+                    'image'=> 'required|mimes:png,jpg,webp,jpeg',
+                ]);
             }
-            $image = $request->file('image');
-            $newName = Str::slug($project->title).'-image-'.Str::random(5).'.'.$image->getClientOriginalExtension();
-            $destination = public_path('images/projects/'.$newName);
-            Image::make($image)->save($destination);
-            $project->featured_image = $newName;
+            $project->title =$request->title;
+            $project->slug = Str::slug($request->title);
+            $project->description =$request->description;
             $project->save();
+            if($request->hasFile('image')){
+                if($project->featured_image){
+                    $oldImage = public_path('images/projects/'.$project->featured_image);
+                    if(file_exists($oldImage)){
+                        unlink($oldImage);
+                    }
+                }
+                $image = $request->file('image');
+                $newName = Str::slug($project->title).'-image-'.Str::random(5).'.'.$image->getClientOriginalExtension();
+                $destination = public_path('images/projects/'.$newName);
+                Image::make($image)->save($destination);
+                $project->featured_image = $newName;
+                $project->save();
+            }
+            return redirect()->route('projects.index')->with('success','Project Updated!');
+
+        }else{
+            return abort(404);
         }
-        return redirect()->route('projects.index')->with('success','Project Updated!');
     }
     public function multipleImageCreate($slug){
         $project = Project::where('slug',$slug)->first();
@@ -140,45 +164,70 @@ class ProjectController extends Controller
      */
     public function destroy(Project $project)
     {
-        $project->delete();
-        return back()->with('success','Project moved to trash');
+        if(auth()->user()->can('project management')){
+            $project->delete();
+            return back()->with('success','Project moved to trash');
+        }else{
+            return abort(404);
+        }
     }
     public function multipleImageUpdate(Request $request)
     {
-
-        if($request->hasFile('gallery')){
-            $size = 0;
-            foreach ($request->file('gallery') as $image) {
-                $ext = $image->getClientOriginalExtension();
-                $valid_ext = ['jpg','jpeg','png','webp'];
-                if(!in_array($ext, $valid_ext)){
-                    return back()->with('validation_error','Uploaded file only suport jpg, jpeg, png or webp');
+        if(auth()->user()->can('project management')){
+            if($request->hasFile('gallery')){
+                $size = 0;
+                foreach ($request->file('gallery') as $image) {
+                    $ext = Str::lower($image->getClientOriginalExtension());
+                    $valid_ext = ['jpg','jpeg','png','webp'];
+                    if(!in_array($ext, $valid_ext)){
+                        return back()->with('validation_error','Uploaded file only suport jpg, jpeg, png or webp');
+                    }
+                    $size += $image->getSize();
                 }
-                $size += $image->getSize();
-            }
-            if($size < 10485760){
-                $loopKey = 0;
-                foreach ($request->file('gallery') as $key => $image) {
-                    $loopKey += $key;
-                    $project = Project::where('slug',$request->project_slug)->first();
-                    $projectGallery = new ProjectImageGallery;
-                    $newName = Str::slug($project->title).'-image-'.Str::random(5).'.'.$image->getClientOriginalExtension();
-                    $path = public_path('images/projects/image_gallery/').$project->slug.'/';
-                    File::makeDirectory($path, $mode = 0777, true, true);
-                    Image::make($image)->save($path.$newName);
-                    $projectGallery->project_id = $project->id;
-                    $projectGallery->image = $newName;
-                    $projectGallery->save();
-                }
-                if($loopKey > 0){
-                    return back()->with('success','New Images Added!');
+                if($size < 10485760){
+                    $loopKey = 0;
+                    foreach ($request->file('gallery') as $key => $image) {
+                        $loopKey += $key;
+                        $project = Project::where('slug',$request->project_slug)->first();
+                        $projectGallery = new ProjectImageGallery;
+                        $newName = Str::slug($project->title).'-image-'.Str::random(5).'.'.$image->getClientOriginalExtension();
+                        $path = public_path('images/projects/image_gallery/').$project->slug.'/';
+                        File::makeDirectory($path, $mode = 0777, true, true);
+                        Image::make($image)->save($path.$newName);
+                        $projectGallery->project_id = $project->id;
+                        $projectGallery->image = $newName;
+                        $projectGallery->save();
+                    }
+                    if($loopKey > 0){
+                        return back()->with('success','New Images Added!');
+                    }else{
+                        return back()->with('success','New Image Added!');
+                    }
                 }else{
-                    return back()->with('success','New Image Added!');
+                    return back()->with('validation_error','Uploaded file can not longer than 10MB');
                 }
-            }else{
-                return back()->with('validation_error','Uploaded file can not longer than 10MB');
-            }
 
+            }else{
+                return back()->with('validation_error','Please choose minimum one image to upload');
+            }
+        }else{
+            return abort(404);
+        }
+    }
+    public function multipleImageDelete(Request $request){
+        if(auth()->user()->can('project management')){
+            $project = Project::find($request->project_id);
+            $image = ProjectImageGallery::find($request->image_id);
+            $oldImage = public_path('images/projects/image_gallery/'.$project->slug.'/'.$image->image);
+            if(file_exists($oldImage)){
+                unlink($oldImage);
+                $image->delete();
+                return back()->with('success','Image deleted');
+            }else{
+                return back()->with('error','Failed to delete image');
+            }
+        }else{
+            return abort(404);
         }
     }
 }
